Fix t_L calculation for work-conserving policies on uniprocessor
t_L (as defined below Eq. 3 in RTSS17) is corrected to use the min
r_x^max of J_x where J_x is an eligible job (priority-eligible and
IIP-eligible) according to Definition 4, instead of r_j^max. This
solves the problem where the processor is sometimes incorrectly
left idle when it should not be according to the work-conserving
policy.